Refactor: move preprocessing into its own module
